+++ UDF/View/AlertBuilder.swift
@@ -24,7 +24,7 @@ public enum AlertBuilder {
 
         public enum Status: Equatable {
             case presented(AlertStyle)
-            case presentedWithStyle(TheAlertStyle)
+            case presentedWithStyle(AlertStyleUpd)
             case dismissed
 
             public static func == (lhs: Self, rhs: Self) -> Bool {
@@ -75,11 +75,6 @@ public enum AlertBuilder {
             status = .presented(style)
         }
         
-        public init(theStyle: TheAlertStyle) {
-            id = theStyle.id
-            status = .presentedWithStyle(theStyle)
-        }
-
         public init<AlertId: Hashable>(id: AlertId) {
             if let builder = AlertBuilder.alertBuilders[id] {
                 self = .init(style: builder())
@@ -87,63 +82,19 @@ public enum AlertBuilder {
                 self = .dismissed
             }
         }
-    }
-    
-    public struct TheAlertStyle: Equatable {
-        public static func == (lhs: Self, rhs: Self) -> Bool {
-            lhs.id == rhs.id
+        
+        public init(theStyle: AlertStyleUpd) {
+            id = theStyle.id
+            status = .presentedWithStyle(theStyle)
         }
         
-        public var id = UUID()
-        public var alertType: TheAlertType = .none
-        public var title: String = ""
-        public var body: String = ""
-        public var message: String = ""
-        public var actions: [AlertAction] = []
-        
-        public enum TheAlertType {
-            case none
-            ///            func alert<A>(LocalizedStringKey, isPresented: Binding<Bool>, actions: () -> A) -> some View
-            ///            Presents an alert when a given condition is true, using a localized string key for the title.
-            case title
-            case error
-            ///            func alert<A, M>(LocalizedStringKey, isPresented: Binding<Bool>, actions: () -> A, message: () -> M) -> some View
-            ///            Presents an alert with a message when a given condition is true, using a localized string key for a title.
-            case message
+        public init(errorStyle: String?) {
+            if let errorStyle, !errorStyle.isEmpty {
+                self = .init(theStyle: .init(error: errorStyle))
+            } else {
+                self = .init()
+            }
         }
-
-        public init() {}
-        
-        public init(error: String) {
-            self.alertType = .error
-            self.title = error
-        }
-        
-        public init(title: String, body: String, actions: [AlertAction]) {
-            self.alertType = .title
-            self.title = title
-            self.body = body
-            self.actions = actions
-        }
-        
-        public init(title: String, body: String, message: String, actions: [AlertAction]) {
-            self.alertType = .message
-            self.title = title
-            self.body = body
-            self.message = message
-            self.actions = actions
-        }
-    }
-    
-    public struct AlertAction: Identifiable {
-        public static func == (lhs: AlertBuilder.AlertAction, rhs: AlertBuilder.AlertAction) -> Bool {
-            return lhs.id == rhs.id
-        }
-        
-        public var id = UUID()
-        public var title: String
-        public var role: ButtonRole?
-        public var action: () -> () = {}
     }
 
     public struct AlertStyle: Equatable {
